# docs/api_swagger.yaml
openapi: 3.0.3
info:
  title: Intelligent Email Agent API
  description: |
    A comprehensive API for intelligent email analysis, categorization, and management.
    
    ## Features
    - **Email Processing**: Connect to IMAP servers and process emails
    - **AI Analysis**: Analyze urgency, generate summaries, and provide follow-up suggestions
    - **Categorization**: Automatically categorize emails by type and priority
    - **Notifications**: Send notifications via Slack, Telegram, and WhatsApp
    - **Voice Summaries**: Generate voice summaries using TTS
    - **Configuration Management**: Manage email settings and VIP contacts
    - **Scheduling**: Automated daily summaries and reminders
    
    ## Authentication
    This API uses email credentials for IMAP access. Ensure your email provider supports IMAP and has appropriate security settings enabled.
    
    ## Rate Limits
    - Email processing: 10 requests per minute
    - Notifications: 50 requests per minute
    - Voice generation: 20 requests per minute
    
  version: 1.0.0
  contact:
    name: Email Agent Support
    email: support@emailagent.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.emailagent.com
    description: Production server
...
paths:
...
  /api/email/process-emails:
    post:
      tags:
        - Email Analysis
      summary: Process emails from IMAP server
      description: |
        Connect to an IMAP server and process emails for analysis, categorization, and storage.
        This endpoint will fetch emails, analyze them using AI, categorize them, and store the results.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailProcessingRequest'
            example:
              email_address: "user@example.com"
              password: "app-password"
              imap_server: "imap.gmail.com"
              imap_port: 993
              use_ssl: true
              limit: 50
      responses:
        '200':
          description: Emails processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailProcessingResponse'
              example:
                total_processed: 25
                total_saved: 25
                errors: []
                processing_time: 15.5
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    EmailProcessingRequest:
      type: object
      required:
        - email_address
        - password
        - imap_server
        - imap_port
        - use_ssl
      properties:
        email_address:
          type: string
          format: email
          description: Email address for IMAP connection
        password:
          type: string
          description: Email password or app password
        imap_server:
          type: string
          description: IMAP server hostname
        imap_port:
          type: integer
          minimum: 1
          maximum: 65535
          description: IMAP server port
        use_ssl:
          type: boolean
          description: Whether to use SSL/TLS connection
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 50
          description: Maximum number of emails to process
    EmailProcessingResponse:
      type: object
      properties:
        total_processed:
          type: integer
          description: Total number of emails processed
        total_saved:
          type: integer
          description: Total number of emails saved to database
        errors:
          type: array
          items:
            type: string
          description: List of processing errors
        processing_time:
          type: number
          format: float
          description: Processing time in seconds
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []
...